Include awaitNextMission as generic idle state
And allow stop event in home state
